blob/requirements.txt
azure-functions
azure-storage-blob
aiohttp